requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.42.0
h11==0.8.1
httptools==0.0.13
idna==2.8
passlib==1.7.1
pycparser==2.19
pydantic==0.32.2
PyJWT==1.7.1
python-dotenv==0.10.3
six==1.13.0
starlette==0.12.9
uvicorn==0.10.3
uvloop==0.14.0rc2
websockets==8.1
